u32 -> u64 patch
(cherry picked from commit ca7a53fe899f53df304bdca9c4a32ce470c84eb2)

+++ yrs/src/block_iter.rs
@@ -10,8 +10,8 @@ use crate::{Assoc, Out, ID};
 #[derive(Debug, Clone)]
 pub(crate) struct BlockIter {
     branch: BranchPtr,
-    index: u32,
-    rel: u32,
+    index: u64,
+    rel: u64,
     next_item: Option<ItemPtr>,
     curr_move: Option<ItemPtr>,
     curr_move_start: Option<ItemPtr>,
@@ -38,7 +38,7 @@ impl BlockIter {
     }
 
     #[inline]
-    pub fn rel(&self) -> u32 {
+    pub fn rel(&self) -> u64 {
         self.rel
     }
 
@@ -70,7 +70,7 @@ impl BlockIter {
         }
     }
 
-    pub fn move_to(&mut self, index: u32, txn: &mut TransactionMut) {
+    pub fn move_to(&mut self, index: u64, txn: &mut TransactionMut) {
         if index > self.index {
             if !self.try_forward(txn, index - self.index) {
                 panic!("Block iter couldn't move forward");
@@ -80,7 +80,7 @@ impl BlockIter {
         }
     }
 
-    fn can_forward(&self, ptr: Option<ItemPtr>, len: u32) -> bool {
+    fn can_forward(&self, ptr: Option<ItemPtr>, len: u64) -> bool {
         if !self.reached_end || self.curr_move.is_some() {
             if len > 0 {
                 return true;
@@ -96,13 +96,13 @@ impl BlockIter {
         false
     }
 
-    pub fn forward<T: ReadTxn>(&mut self, txn: &T, len: u32) {
+    pub fn forward<T: ReadTxn>(&mut self, txn: &T, len: u64) {
         if !self.try_forward(txn, len) {
             panic!("Length exceeded")
         }
     }
 
-    pub fn try_forward<T: ReadTxn>(&mut self, txn: &T, mut len: u32) -> bool {
+    pub fn try_forward<T: ReadTxn>(&mut self, txn: &T, mut len: u64) -> bool {
         if len == 0 && self.next_item.is_none() {
             return true;
         }
@@ -183,7 +183,7 @@ impl BlockIter {
         }
     }
 
-    pub fn backward<T: ReadTxn>(&mut self, txn: &mut T, mut len: u32) {
+    pub fn backward<T: ReadTxn>(&mut self, txn: &mut T, mut len: u64) {
         if self.index < len {
             panic!("Length exceeded");
         }
@@ -297,7 +297,7 @@ impl BlockIter {
         self.reached_end = false;
     }
 
-    pub fn delete(&mut self, txn: &mut TransactionMut, mut len: u32) {
+    pub fn delete(&mut self, txn: &mut TransactionMut, mut len: u64) {
         let mut item = self.next_item;
         if self.index + len > self.branch.content_len() {
             panic!("Length exceeded");
@@ -358,15 +358,15 @@ impl BlockIter {
         self.next_item = item;
     }
 
-    pub(crate) fn slice<T: ReadTxn>(&mut self, txn: &T, buf: &mut [Out]) -> u32 {
-        let mut len = buf.len() as u32;
+    pub(crate) fn slice<T: ReadTxn>(&mut self, txn: &T, buf: &mut [Out]) -> u64 {
+        let mut len = buf.len() as u64;
         if self.index + len > self.branch.content_len() {
             return 0;
         }
         self.index += len;
         let mut next_item = self.next_item;
         let encoding = txn.store().options.offset_kind;
-        let mut read = 0u32;
+        let mut read = 0u64;
         while len > 0 {
             if !self.reached_end {
                 while let Some(item) = next_item {
@@ -380,7 +380,7 @@ impl BlockIter {
                             let r = item
                                 .content
                                 .read(self.rel as usize, &mut buf[read as usize..])
-                                as u32;
+                                as u64;
                             read += r;
                             len -= r;
                             if self.rel + r == item.content_len(encoding) {
